+++ src/TestSuite/Fixture.php
@@ -79,6 +79,7 @@ class Fixture
      */
     public function initialize()
     {
+       
     }
     
     /**