+++ tests/TestCase/Utility/DateTest.php
@@ -56,29 +56,47 @@ class DateTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals('19/03/2019', Date::format('2019-03-19 11:54:00','d/m/Y'));
     }
 
+    public function testFormatFormat(){
+        // Test No TZ
+        Date::locale(['date'=>'d/m/Y','datetime' => 'd/m/Y g:i A','time' =>'g:i A']);
+        $this->assertEquals('11-06-2019 4:00 PM', Date::format('2019-06-11 16:00:00','d-m-Y g:i A'));
+
+        Date::locale(['date'=>'d/m/Y','datetime' => 'd/m/Y g:i A','time' =>'g:i A','timezone' => 'Europe/Madrid']);
+        $this->assertEquals('11-06-2019 6:00 PM', Date::format('2019-06-11 16:00:00','d-m-Y g:i A'));
+    }
+
+    public function testFormatAutoDetect(){
+        Date::locale(['date'=>'d/m/Y','datetime' => 'd/m/Y g:i A','time' =>'g:i A']);
+        $this->assertEquals('11/06/2019 4:00 PM', Date::format('2019-06-11 16:00:00'));
+        $this->assertEquals('11/06/2019', Date::format('2019-06-11'));
+        $this->assertEquals('4:00 PM', Date::format('16:00:00'));
+    }
+
     public function testFormatDate()
     {
-        Date::locale(['date' => 'd/m/Y']);
-        $this->assertEquals('19/03/2019', Date::format('2019-03-19'));
+        Date::locale(['date'=>'d/m/Y','datetime' => 'd/m/Y g:i A']);
+        $this->assertEquals('19/03/2019', Date::formatDate('2019-03-19'));
+
+        Date::locale(['date'=>'d/m/Y','datetime' => 'd/m/Y g:i A','timezone' => 'Europe/Madrid']);
+        $this->assertEquals('19/03/2019', Date::formatDate('2019-03-19'));
     }
 
     public function testFormatDateTime()
     {
         Date::locale(['datetime' => 'd/m/Y g:i A']);
-        $this->assertEquals('19/03/2019 11:54 AM', Date::format('2019-03-19 11:54:00'));
+        $this->assertEquals('19/03/2019 11:54 AM', Date::formatDateTime('2019-03-19 11:54:00'));
+
         Date::locale(['datetime' => 'd/m/Y g:i A','timezone' => 'Europe/Madrid']);
-        $this->assertEquals('19/03/2019 12:54 PM', Date::format('2019-03-19 11:54:00'));
+        $this->assertEquals('19/03/2019 12:54 PM', Date::formatDateTime('2019-03-19 11:54:00'));
     }
 
     public function testFormatTime()
     {
         Date::locale(['time' => 'g:i A']);
-        $this->assertEquals('11:54 AM', Date::format('11:54:00'));
-        Date::locale(['time' => 'g:i A','timezone' => 'Europe/Madrid']);
-        $this->assertEquals('11:54 AM', Date::format('11:54:00'));
+        $this->assertEquals('11:54 AM', Date::formatTime('11:54:00'));
 
         Date::locale(['time' => 'g:i A','timezone' => 'Europe/Madrid']);
-        $this->assertEquals('12:54 PM', Date::formatTime('2019-01-01 11:54:00'));
+        $this->assertEquals('11:54 AM', Date::formatTime('11:54:00'));
     }
 
     public function testParseDate()